# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CeetFeedback API",
        "description": "API documentation for CeetFeedback",
        "contact": {
            "email": "blezanatanael@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "V1"
    },
    "paths": {
...
        "/api/v1/phases": {
            "get": {
                "tags": [
                    "Phases"
                ],
                "summary": "Récupérer la liste des phases",
                "operationId": "ad9c32ec85ef06c7a227b0fa5bfaf507",
                "responses": {
                    "200": {
                        "description": "Liste des phases récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PhaseSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Phases"
                ],
                "summary": "Créer une nouvelle phase",
                "operationId": "2ac6f56404c6e505894c1d5bbd163a38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhaseSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Phase créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Phase créée avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PhaseSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/phases/{id}": {
            "put": {
                "tags": [
                    "Phases"
                ],
                "summary": "Mettre à jour une phase",
                "operationId": "ae9c9e9ea6fe56b7614021f319ffca5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhaseSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Phase mise à jour avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Phases"
                ],
                "summary": "Supprimer une phase (soft delete)",
                "operationId": "18f54534c9f8c3a0dc79f50bc6b5b338",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Phase supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PhaseSchema": {
                "title": "Phase",
                "description": "Ressource phase",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "phase-abc123"
                    },
                    "name": {
                        "type": "string",
                        "example": "Phase de conception"
                    },
                    "description": {
                        "type": "string",
                        "example": "Cette phase couvre la planification et la conception du projet."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-18T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}